Move the valid_* functions into engine/check.py
